# docs/swagger.yaml
definitions:
  models.Post:
    properties:
      content:
        type: string
      id:
        type: integer
      post_images:
        items:
          $ref: '#/definitions/models.PostImage'
        type: array
      published_at:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      user_id:
        type: integer
      vote_count:
        type: integer
      votes:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.PostImage:
    properties:
      id:
        type: integer
      post_id:
        type: integer
      url:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
    type: object
...
  models.Tag:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
...
  requests.StoreUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: integer
    required:
    - email
    - name
    - password
    - role_id
    type: object
...
  requests.UpdateUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  results.IDResult:
    properties:
      id:
        type: integer
    type: object
  results.JSONResult:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  results.UserResult:
    properties:
      user:
        $ref: '#/definitions/models.User'
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /api/v1/posts/{id}/votes/down:
    get:
      description: Downvote a post by inserting ID in route param, you can only do
        it once per post.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.IDResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Downvote a post
      tags:
      - Post Votes
  /api/v1/posts/{id}/votes/up:
    get:
      description: Upvote a post by inserting ID in route param, you can only do it
        once per post.
      parameters:
      - description: post id
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.IDResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Upvote a post
      tags:
      - Post Votes
...
  /api/v1/users:
    get:
      description: Get a list of users, this endpoint only for admin.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/results.UserResult'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get all users (admin only)
      tags:
      - Users
    post:
      description: Create new user, this endpoint only for admin.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.StoreUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.UserResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Create new user (admin only)
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Delete user by ID, you can only delete your own data.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.IDResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete user by ID
      tags:
      - Users
    get:
      description: Get an existing user byID.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.UserResult'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get user by ID
      tags:
      - Users
    put:
      description: Update user by ID, you can only update your own data.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  $ref: '#/definitions/results.UserResult'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  items:
                    additionalProperties:
                      type: string
                    type: object
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/results.JSONResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Update user by ID
      tags:
      - Users
swagger: "2.0"